requirements.txt
Django>=3.2
psycopg2-binary>=2.8
django-bootstrap-v5
djangorestframework
requests
openpyxl
coverage
django-cors-headers
django-rest-knox
drf-spectacular